Split server.py — push + desktop route domains (#586)
Extract the push (3 handlers) and desktop (11 handlers) route domains from
server.py into routes/push.py and routes/desktop.py, mirroring the #584
mixin/registrar seam. Pure relocation — route set invariant.

- PushRoutesMixin: api_push_config/subscribe/unsubscribe + register_push_routes.
  _fanout_push STAYS on the base class (shared by desktop notifications + the
  service watchdog), not part of this domain's HTTP surface.
- DesktopRoutesMixin: api_desktop_windows/open/close/focus/tile/minimize_all/
  collage/layout/notification/notification_dismiss/notifications_list +
  register_desktop_routes. active_notifications + broadcast_dashboard +
  _fanout_push stay on base, resolved via MRO.

Verify: route-snapshot guard green (150-route freeze), full suite
(2147 passed, 8 skipped), portal boots + push/desktop routes serve 200.

Closes #586

Built by dotdev.dev
